moved files to folders and fix some errors
