+++ src/App.js
@@ -1,11 +1,11 @@
 // import logo from './logo.svg';
 import './App.css';
-import LoginSignup from './components/LoginSignup/LoginSignup'
+import LoginSignup from './LoginSignup';
 
 function App() {
   return (
     <div>
-      <LoginSignup/>
+      <LoginSignup />
     </div>
   );
 }
